重构 PDF 文本抽取共享逻辑，并补齐 styles.css
- 抽取 pageItemsToText 与 loadPdfPages 到 src/utils/pdf/pdfPages.ts，vault 索引与 chat 上传链路共用
- 重新生成 styles.css，纳入新增的 .smtcmp-chat-user-input-upload-button 样式
